Added Dynamic Spawn Slots
Added Dynamic spawns. Ran out of time to properly test, will revise tomorrow! This should demonstrate the ease of use in adding dyn-spawns with pre-loaded waypoints. Be on the lookout for hacked payloads! (Including some whacky ones for the F-4E!)

Farps will be off-center, but are easily tweaked. Please change the Moose pathing to your own to run!